iqm_station_control_client/requirements/base.txt
iqm-exa-common>=26,<27
iqm-data-definitions >= 2.13, < 3.0
opentelemetry-exporter-otlp == 1.25.0
protobuf >= 4.25.3, < 5.0
types-protobuf
grpcio >= 1.65.4, < 2.0
pydantic >= 2.10.4, <3.0
PyYAML == 6.0
requests == 2.32.3
types-requests
tqdm >= 4.59.0
types-tqdm